ci: add basic boilerplate for invoking unit tests
This change adds a basic entrypoint that is intended for invoking unit
tests. It is initially implemented as a no-op to ensure that the new
prowjob does not cause any disruption (for example, if another PR is
merged with broken unit tests).

Once the prowjob is created, a subsequent PR will update this script to
actually invoke the unit tests.